xml.cpp: bunch of size_t where ssize_t probably should've been used.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@18492 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/xml/xml.cpp
@@ -268,7 +268,7 @@ int File::read (char *aBuf, int aLen)
     size_t len = aLen;
     int vrc = RTFileRead (m->handle, aBuf, len, &len);
     if (RT_SUCCESS (vrc))
-        return len;
+        return (int)len;
 
     throw EIPRTFailure (vrc);
 }
@@ -278,7 +278,7 @@ int File::write (const char *aBuf, int aLen)
     size_t len = aLen;
     int vrc = RTFileWrite (m->handle, aBuf, len, &len);
     if (RT_SUCCESS (vrc))
-        return len;
+        return (int)len;
 
     throw EIPRTFailure (vrc);
 
@@ -359,7 +359,7 @@ int MemoryBuf::read (char *aBuf, int aLen)
     memcpy (aBuf, m->buf + m->pos, len);
     m->pos += len;
 
-    return len;
+    return (int)len;
 }
 
 ////////////////////////////////////////////////////////////////////////////////
